Reorganized the code in filtering directory
No functional changes.
