+++ docs/_templates/autosummary/module.rst
@@ -1,5 +1,14 @@
-{% if fullname == 'zrad.filtering.filtering_definitions' %}
-Filter Definitions
+{% if fullname == 'zrad.filtering.base' %}
+Filtering Base
+==============
+{% elif fullname == 'zrad.filtering.factory' %}
+Filter Factory
+==============
+{% elif fullname == 'zrad.filtering.spatial' %}
+Spatial Filters
+===============
+{% elif fullname == 'zrad.filtering.wavelet' %}
+Wavelet Filters
 ==================
 {% elif fullname == 'zrad.radiomics.radiomics_definitions' %}
 Radiomics Definitions
