+++ zrad/filtering/__init__.py
@@ -1,2 +1,15 @@
-from .filtering import create_filter, Filtering
-from .filtering_definitions import BaseFilter, Mean, LoG, Laws, Gabor, Wavelets2D, Wavelets3D
+from .base import BaseFilter
+from .factory import create_filter
+from .spatial import Mean, LoG, Laws, Gabor
+from .wavelet import Wavelets2D, Wavelets3D
+
+__all__ = [
+    'BaseFilter',
+    'create_filter',
+    'Mean',
+    'LoG',
+    'Laws',
+    'Gabor',
+    'Wavelets2D',
+    'Wavelets3D',
+]
